backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
email-validator==2.1.0.post1
python-multipart==0.0.6

# Feeds parsing
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
redis==5.0.1
